Removed a fair amount of code duplication
