+++ src/sensors/softfusion/drivers/icm45686.h
@@ -23,9 +23,7 @@
 
 #pragma once
 
-#include <algorithm>
-#include <array>
-#include <cstdint>
+#include "icm45base.h"
 
 namespace SlimeVR::Sensors::SoftFusion::Drivers {
 
@@ -37,39 +35,14 @@ namespace SlimeVR::Sensors::SoftFusion::Drivers {
 // Timestamps reading not used, as they're useless (constant predefined increment)
 
 template <typename I2CImpl>
-struct ICM45686 {
-	static constexpr uint8_t Address = 0x68;
+struct ICM45686 : ICM45Base<I2CImpl> {
 	static constexpr auto Name = "ICM-45688";
 	static constexpr auto Type = ImuID::ICM45686;
 
-	static constexpr float GyrTs = 1.0 / 409.6;
-	static constexpr float AccTs = 1.0 / 102.4;
-
-	static constexpr float MagTs = 1.0 / 100;
-
-	static constexpr float GyroSensitivity = 131.072f;
-	static constexpr float AccelSensitivity = 16384.0f;
-
-	static constexpr bool Uses32BitSensorData = true;
-
-	I2CImpl i2c;
-	SlimeVR::Logging::Logger& logger;
 	ICM45686(I2CImpl i2c, SlimeVR::Logging::Logger& logger)
-		: i2c(i2c)
-		, logger(logger) {}
+		: ICM45Base<I2CImpl>(i2c, logger) {}
 
-	struct Regs {
-		struct WhoAmI {
-			static constexpr uint8_t reg = 0x72;
-			static constexpr uint8_t value = 0xe9;
-		};
-		static constexpr uint8_t TempData = 0x0c;
-
-		struct DeviceConfig {
-			static constexpr uint8_t reg = 0x7f;
-			static constexpr uint8_t valueSwReset = 0b11;
-		};
-
+	struct AdditionalRegs {
 		struct Pin9Config {
 			static constexpr uint8_t reg = 0x31;
 			static constexpr uint8_t value = 0b00000110;  // pin 9 to clkin
@@ -79,126 +52,15 @@ struct ICM45686 {
 			static constexpr uint8_t reg = 0x26;
 			static constexpr uint8_t value = 0b00100011;  // enable RTC
 		};
-
-		struct GyroConfig {
-			static constexpr uint8_t reg = 0x1c;
-			static constexpr uint8_t value
-				= (0b0000 << 4) | 0b0111;  // 4000dps, odr=409.6Hz
-		};
-
-		struct AccelConfig {
-			static constexpr uint8_t reg = 0x1b;
-			static constexpr uint8_t value
-				= (0b000 << 4) | 0b1001;  // 32g, odr = 102.4Hz
-		};
-
-		struct FifoConfig0 {
-			static constexpr uint8_t reg = 0x1d;
-			static constexpr uint8_t value
-				= (0b01 << 6) | (0b011111);  // stream to FIFO mode, FIFO depth
-											 // 8k bytes <-- this disables all APEX
-											 // features, but we don't need them
-		};
-
-		struct FifoConfig3 {
-			static constexpr uint8_t reg = 0x21;
-			static constexpr uint8_t value = (0b1 << 0) | (0b1 << 1) | (0b1 << 2)
-										   | (0b1 << 3);  // enable FIFO,
-														  // enable accel,
-														  // enable gyro,
-														  // enable hires mode
-		};
-
-		struct PwrMgmt0 {
-			static constexpr uint8_t reg = 0x10;
-			static constexpr uint8_t value
-				= 0b11 | (0b11 << 2);  // accel in low noise mode, gyro in low noise
-		};
-
-		static constexpr uint8_t FifoCount = 0x12;
-		static constexpr uint8_t FifoData = 0x14;
 	};
 
-#pragma pack(push, 1)
-	struct FifoEntryAligned {
-		union {
-			struct {
-				int16_t accel[3];
-				int16_t gyro[3];
-				uint16_t temp;
-				uint16_t timestamp;
-				uint8_t lsb[3];
-			} part;
-			uint8_t raw[19];
-		};
-	};
-#pragma pack(pop)
-
-	static constexpr size_t FullFifoEntrySize = sizeof(FifoEntryAligned) + 1;
+	using ICM45Base<I2CImpl>::i2c;
 
 	bool initialize() {
-		// perform initialization step
-		i2c.writeReg(Regs::DeviceConfig::reg, Regs::DeviceConfig::valueSwReset);
-		delay(35);
-		i2c.writeReg(Regs::Pin9Config::reg, Regs::Pin9Config::value);
-		i2c.writeReg(Regs::RtcConfig::reg, Regs::RtcConfig::value);
-		i2c.writeReg(Regs::GyroConfig::reg, Regs::GyroConfig::value);
-		i2c.writeReg(Regs::AccelConfig::reg, Regs::AccelConfig::value);
-		i2c.writeReg(Regs::FifoConfig0::reg, Regs::FifoConfig0::value);
-		i2c.writeReg(Regs::FifoConfig3::reg, Regs::FifoConfig3::value);
-		i2c.writeReg(Regs::PwrMgmt0::reg, Regs::PwrMgmt0::value);
-		delay(1);
-
-		return true;
-	}
-
-	float getDirectTemp() const {
-		const auto value = static_cast<int16_t>(i2c.readReg16(Regs::TempData));
-		float result = ((float)value / 132.48f) + 25.0f;
-		return result;
-	}
-
-	template <typename AccelCall, typename GyroCall>
-	void bulkRead(AccelCall&& processAccelSample, GyroCall&& processGyroSample) {
-		const auto fifo_packets = i2c.readReg16(Regs::FifoCount);
-		const auto fifo_bytes = fifo_packets * sizeof(FullFifoEntrySize);
-
-		std::array<uint8_t, FullFifoEntrySize * 8> read_buffer;  // max 8 readings
-		const auto bytes_to_read = std::min(
-									   static_cast<size_t>(read_buffer.size()),
-									   static_cast<size_t>(fifo_bytes)
-								   )
-								 / FullFifoEntrySize * FullFifoEntrySize;
-		i2c.readBytes(Regs::FifoData, bytes_to_read, read_buffer.data());
-		for (auto i = 0u; i < bytes_to_read; i += FullFifoEntrySize) {
-			FifoEntryAligned entry;
-			memcpy(
-				entry.raw,
-				&read_buffer[i + 0x1],
-				sizeof(FifoEntryAligned)
-			);  // skip fifo header
-			const int32_t gyroData[3]{
-				static_cast<int32_t>(entry.part.gyro[0]) << 4
-					| (entry.part.lsb[0] & 0xf),
-				static_cast<int32_t>(entry.part.gyro[1]) << 4
-					| (entry.part.lsb[1] & 0xf),
-				static_cast<int32_t>(entry.part.gyro[2]) << 4
-					| (entry.part.lsb[2] & 0xf),
-			};
-			processGyroSample(gyroData, GyrTs);
-
-			if (entry.part.accel[0] != -32768) {
-				const int32_t accelData[3]{
-					static_cast<int32_t>(entry.part.accel[0]) << 4
-						| (static_cast<int32_t>(entry.part.lsb[0]) & 0xf0 >> 4),
-					static_cast<int32_t>(entry.part.accel[1]) << 4
-						| (static_cast<int32_t>(entry.part.lsb[1]) & 0xf0 >> 4),
-					static_cast<int32_t>(entry.part.accel[2]) << 4
-						| (static_cast<int32_t>(entry.part.lsb[2]) & 0xf0 >> 4),
-				};
-				processAccelSample(accelData, AccTs);
-			}
-		}
+		ICM45Base<I2CImpl>::softResetIMU();
+		i2c.writeReg(AdditionalRegs::Pin9Config::reg, AdditionalRegs::Pin9Config::value);
+		i2c.writeReg(AdditionalRegs::RtcConfig::reg, AdditionalRegs::RtcConfig::value);
+		return ICM45Base<I2CImpl>::initializeBase();
 	}
 };
